Moved creation of the amici objective function to a separate class, leaving the importer really independent of amici
